pyproject.toml
[project]
name = "pytorch-mlflow-sample"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "line-bot-sdk>=3.17.1",
    "mlflow>=2.22.0",
    "numpy>=2.2.5",
    "pydantic>=2.11.3",
    "python-dotenv>=1.1.0",
    "torch>=2.7.0",
    "torchvision>=0.22.0",
    "tqdm>=4.67.1",
    "typer>=0.15.3",
]
